// frontend/src/types/generated/ClientMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SubscriptionFilter } from "./SubscriptionFilter";

export type ClientMessage = { "type": "subscribe", filter: SubscriptionFilter | null, } | { "type": "unsubscribe" } | { "type": "ping" };