// frontend/src/types/generated/ServerMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventEnvelope } from "./EventEnvelope";
import type { SubscriptionFilter } from "./SubscriptionFilter";

export type ServerMessage = { "type": "event", envelope: EventEnvelope, } | { "type": "subscribed", filter: SubscriptionFilter | null, } | { "type": "unsubscribed" } | { "type": "pong" } | { "type": "error", message: string, };