// frontend/src/types/generated/EventEnvelope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Event } from "./Event";

/**
 * Envelope wrapping all events with metadata
 */
export type EventEnvelope = { 
/**
 * Unique event ID
 */
id: string, 
/**
 * When the event occurred
 */
timestamp: string, 
/**
 * The actual event
 */
event: Event, };